+++ e2e/e2etests/e2etests.go
@@ -61,6 +61,7 @@ const (
 	 * Solana tests
 	 */
 	TestSolanaDepositName                                 = "solana_deposit"
+	TestSolanaDepositThroughProgramName                   = "solana_deposit_through_program"
 	TestSolanaWithdrawName                                = "solana_withdraw"
 	TestSolanaWithdrawAndCallName                         = "solana_withdraw_and_call"
 	TestZEVMToSolanaCallName                              = "zevm_to_solana_call"
@@ -342,6 +343,7 @@ var AllE2ETests = []runner.E2ETest{
 		"withdraw Ether from ZEVM call a contract",
 		[]runner.ArgDefinition{
 			{Description: "amount in wei", DefaultValue: "100000"},
+			{Description: "gas limit for withdraw", DefaultValue: "250000"},
 		},
 		TestETHWithdrawAndCall,
 	),
@@ -350,6 +352,7 @@ var AllE2ETests = []runner.E2ETest{
 		"withdraw Ether from ZEVM call a contract with no message content",
 		[]runner.ArgDefinition{
 			{Description: "amount in wei", DefaultValue: "100000"},
+			{Description: "gas limit for withdraw", DefaultValue: "250000"},
 		},
 		TestETHWithdrawAndCallNoMessage,
 	),
@@ -382,6 +385,7 @@ var AllE2ETests = []runner.E2ETest{
 		"withdraw Ether from ZEVM, revert, then abort with onAbort, check onAbort can created cctx",
 		[]runner.ArgDefinition{
 			{Description: "amount in wei", DefaultValue: "1000000000000000000"},
+			{Description: "gas limit for withdraw", DefaultValue: "250000"},
 		},
 		TestETHWithdrawRevertAndAbort,
 		runner.WithMinimumVersion("v29.0.0"),
@@ -569,6 +573,14 @@ var AllE2ETests = []runner.E2ETest{
 			{Description: "amount in lamport", DefaultValue: "24000000"},
 		},
 		TestSolanaDeposit,
+	),
+	runner.NewE2ETest(
+		TestSolanaDepositThroughProgramName,
+		"deposit SOL into ZEVM through example connected program",
+		[]runner.ArgDefinition{
+			{Description: "amount in lamport", DefaultValue: "24000000"},
+		},
+		TestSolanaDepositThroughProgram,
 	),
 	runner.NewE2ETest(
 		TestSolanaWithdrawName,