+++ zetaclient/chains/bitcoin/client/helpers.go
@@ -3,6 +3,8 @@ package client
 import (
 	"context"
 	"fmt"
+	"math"
+	"strings"
 	"time"
 
 	types "github.com/btcsuite/btcd/btcjson"
@@ -17,9 +19,22 @@ const (
 	// FeeRateRegnet is the hardcoded fee rate for regnet
 	FeeRateRegnet = 1
 
+	// FeeRateRegnetRBF is the hardcoded fee rate for regnet RBF
+	// The zetacore bumps CCTX's fee rate every 40 minutes, and we can't wait that long in the E2E test.
+	// For simplicity, zetaclient uses a constant fee rate (> above 1 sat/vB) to test RBF in the regnet.
+	FeeRateRegnetRBF = 5
+
 	// maxBTCSupply is the maximum supply of Bitcoin
 	maxBTCSupply = 21000000.0
 )
+
+// MempoolTxsAndFees contains the information of pending mempool txs and fees
+type MempoolTxsAndFees struct {
+	TotalTxs   int64
+	TotalFees  int64
+	TotalVSize int64
+	AvgFeeRate uint64
+}
 
 // GetBlockVerboseByStr alias for GetBlockVerbose
 func (c *Client) GetBlockVerboseByStr(ctx context.Context, blockHash string) (*types.GetBlockVerboseTxResult, error) {
@@ -115,7 +130,7 @@ func (c *Client) GetRawTransactionResult(ctx context.Context,
 }
 
 // GetEstimatedFeeRate gets estimated smart fee rate (sat/vB) targeting given block confirmation
-func (c *Client) GetEstimatedFeeRate(ctx context.Context, confTarget int64) (satsPerByte int64, err error) {
+func (c *Client) GetEstimatedFeeRate(ctx context.Context, confTarget int64) (satsPerByte uint64, err error) {
 	// RPC 'EstimateSmartFee' is not available in regnet
 	if c.isRegnet {
 		return FeeRateRegnet, nil
@@ -135,7 +150,13 @@ func (c *Client) GetEstimatedFeeRate(ctx context.Context, confTarget int64) (sat
 	if feeRate <= 0 || feeRate >= maxBTCSupply {
 		return 0, fmt.Errorf("invalid fee rate: %f", feeRate)
 	}
-	return common.FeeRateToSatPerByte(feeRate), nil
+
+	feeRateUint, err := common.FeeRateToSatPerByte(feeRate)
+	if err != nil {
+		return 0, errors.Wrapf(err, "invalid fee rate: %f", feeRate)
+	}
+
+	return feeRateUint, nil
 }
 
 // GetTransactionFeeAndRate gets the transaction fee and rate for a given tx result
@@ -207,6 +228,81 @@ func (c *Client) Healthcheck(ctx context.Context) (time.Time, error) {
 	}
 
 	return header.Timestamp, nil
+}
+
+// GetMempoolTxsAndFees returns the information of all pending parent txs and fees of a given tx (inclusive)
+//
+// A parent tx is defined as:
+//   - a tx that is also pending in the mempool
+//   - a tx that has its first output spent by the child as first input
+func (c *Client) GetMempoolTxsAndFees(
+	ctx context.Context,
+	childHash string,
+) (txsAndFees MempoolTxsAndFees, err error) {
+	totalFeesFloat := float64(0)
+
+	// loop through all parents
+	parentHash := childHash
+	for {
+		memplEntry, err := c.GetMempoolEntry(ctx, parentHash)
+		if err != nil {
+			if isTxNotInMempoolError(err) {
+				// not a mempool tx, stop looking for parents
+				break
+			}
+			return txsAndFees, errors.Wrapf(err, "unable to get mempool entry for tx %s", parentHash)
+		}
+
+		// accumulate fees and vsize
+		txsAndFees.TotalTxs++
+		totalFeesFloat += memplEntry.Fee
+		txsAndFees.TotalVSize += int64(memplEntry.VSize)
+
+		// find the parent tx
+		tx, err := c.GetRawTransactionByStr(ctx, parentHash)
+		if err != nil {
+			return txsAndFees, errors.Wrapf(err, "unable to get tx %s", parentHash)
+		}
+		parentHash = tx.MsgTx().TxIn[0].PreviousOutPoint.Hash.String()
+
+		// check timeout to avoid infinite loop
+		if deadline, ok := ctx.Deadline(); ok {
+			if time.Now().After(deadline) {
+				return txsAndFees, errors.Errorf("timeout reached on %dth tx: %s", txsAndFees.TotalTxs, parentHash)
+			}
+		}
+	}
+
+	// no pending tx found
+	if txsAndFees.TotalTxs == 0 {
+		return txsAndFees, errors.Errorf("given tx is not pending: %s", childHash)
+	}
+
+	// convert total fees to satoshis
+	txsAndFees.TotalFees, err = common.GetSatoshis(totalFeesFloat)
+	if err != nil {
+		return txsAndFees, errors.Wrapf(err, "invalid total fees: %f", totalFeesFloat)
+	}
+
+	// sanity check, should never happen
+	if txsAndFees.TotalVSize <= 0 {
+		return txsAndFees, errors.Errorf("invalid totalVSize %d", txsAndFees.TotalVSize)
+	}
+
+	// calculate the average fee rate
+	// #nosec G115 always positive
+	txsAndFees.AvgFeeRate = uint64(math.Ceil(totalFeesFloat / float64(txsAndFees.TotalVSize)))
+
+	return txsAndFees, nil
+}
+
+// isTxNotInMempoolError checks if the given error is due to the transaction not being in the mempool.
+func isTxNotInMempoolError(err error) bool {
+	if err == nil {
+		return false
+	}
+
+	return strings.Contains(err.Error(), "Transaction not in mempool")
 }
 
 func strToHash(s string) (*chainhash.Hash, error) {